backend/requirements.txt
fastapi==0.104.0
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
pillow==10.1.0
numpy==1.26.2
opencv-python==4.8.1.78
requests==2.31.0
pydantic==2.5.0
gunicorn==21.2.0
pytest==7.4.3
httpx==0.25.1
